Fix CMSG-handling Bug
There was a bug where enough _capacity_ was allocated for cmsg's but not
enough actual space.

+++ src/parameters.rs
@@ -293,7 +293,7 @@ impl TestParameter for TtlTestParameter {
                 std::io::ErrorKind::ConnectionRefused,
             ))
         })?;
-        unsafe { Ok(libc::CMSG_SPACE(mem::size_of::<u32>() as libc::c_uint) as usize) }
+        unsafe { Ok(libc::CMSG_SPACE(mem::size_of::<u8>() as libc::c_uint) as usize) }
     }
 
     fn argument_from(&self, cmsg: &ControlMessageOwned, logger: Logger) -> Option<TestArgument> {
